Implement DabContext and refactor SchemaDesigner components to utilize new context for DAB functionality
