+++ extensions/mssql/src/reactviews/pages/SchemaDesigner/schemaDesignerStateProvider.tsx
@@ -5,7 +5,6 @@
 
 import { createContext, useEffect, useRef, useState, useCallback } from "react";
 import { SchemaDesigner } from "../../../sharedInterfaces/schemaDesigner";
-import { Dab } from "../../../sharedInterfaces/dab";
 import { useVscodeWebview } from "../../common/vscodeWebviewProvider";
 import { getCoreRPCs, getErrorMessage } from "../../common/utils";
 import { WebviewRpc } from "../../common/rpc";
@@ -15,10 +14,9 @@ import { flowUtils, foreignKeyUtils } from "./schemaDesignerUtils";
 import eventBus from "./schemaDesignerEvents";
 import {
     registerSchemaDesignerApplyEditsHandler,
-    registerSchemaDesignerDabToolHandlers,
     registerSchemaDesignerGetSchemaStateHandler,
 } from "./schemaDesignerRpcHandlers";
-import { ApiStatus, CoreRPCs } from "../../../sharedInterfaces/webview";
+import { CoreRPCs } from "../../../sharedInterfaces/webview";
 import { filterDeletedEdges, filterDeletedNodes } from "./diff/deletedVisualUtils";
 import {
     applyColumnRenamesToIncomingForeignKeyEdges,
@@ -78,35 +76,7 @@ export interface SchemaDesignerContextProps extends CoreRPCs {
     baselineRevision: number;
     schemaRevision: number;
     notifySchemaChanged: () => void;
-
-    // DAB (Data API Builder) state
-    dabConfig: Dab.DabConfig | null;
     isDabEnabled: () => boolean;
-    initializeDabConfig: () => void;
-    syncDabConfigWithSchema: () => void;
-    updateDabApiTypes: (apiTypes: Dab.ApiType[]) => void;
-    toggleDabEntity: (entityId: string, isEnabled: boolean) => void;
-    toggleDabEntityAction: (entityId: string, action: Dab.EntityAction, isEnabled: boolean) => void;
-    updateDabEntitySettings: (entityId: string, settings: Dab.EntityAdvancedSettings) => void;
-    dabSchemaFilter: string[];
-    setDabSchemaFilter: (schemas: string[]) => void;
-    dabConfigContent: string;
-    dabConfigRequestId: number;
-    generateDabConfig: () => Promise<void>;
-    openDabConfigInEditor: (configContent: string) => void;
-    // DAB Deployment state
-    dabDeploymentState: Dab.DabDeploymentState;
-    openDabDeploymentDialog: () => void;
-    closeDabDeploymentDialog: () => void;
-    setDabDeploymentDialogStep: (step: Dab.DabDeploymentDialogStep) => void;
-    updateDabDeploymentParams: (params: Partial<Dab.DabDeploymentParams>) => void;
-    validateDabDeploymentParams: (
-        containerName: string,
-        port: number,
-    ) => Promise<Dab.ValidateDeploymentParamsResponse>;
-    runDabDeploymentStep: (step: Dab.DabDeploymentStepOrder) => Promise<void>;
-    resetDabDeploymentState: () => void;
-    retryDabDeploymentSteps: () => void;
 }
 
 const SchemaDesignerContext = createContext<SchemaDesignerContextProps>(
@@ -141,21 +111,6 @@ const SchemaDesignerStateProvider: React.FC<SchemaDesignerProviderProps> = ({ ch
     const baselineSchemaRef = useRef<SchemaDesigner.Schema | undefined>(undefined);
     const baselineDefinitionRef = useRef<string | undefined>(undefined);
 
-    // DAB state
-    const [dabConfig, setDabConfig] = useState<Dab.DabConfig | null>(null);
-    const dabConfigRef = useRef<Dab.DabConfig | null>(dabConfig);
-    const extractSchemaRef = useRef<() => SchemaDesigner.Schema>(() => ({ tables: [] }));
-    const [dabSchemaFilter, setDabSchemaFilter] = useState<string[]>([]);
-    const [dabConfigContent, setDabConfigContent] = useState<string>("");
-    const [dabConfigRequestId, setDabConfigRequestId] = useState<number>(0);
-    const [dabDeploymentState, setDabDeploymentState] = useState<Dab.DabDeploymentState>(
-        Dab.createDefaultDeploymentState(),
-    );
-
-    useEffect(() => {
-        dabConfigRef.current = dabConfig;
-    }, [dabConfig]);
-
     const { onPushUndoState, maybeAutoArrangeForToolBatch } = useSchemaDesignerToolBatchHandlers({
         reactFlow,
         resetView,
@@ -172,10 +127,6 @@ const SchemaDesignerStateProvider: React.FC<SchemaDesignerProviderProps> = ({ ch
     const notifySchemaChanged = useCallback(() => {
         setSchemaRevision((revision) => revision + 1);
     }, []);
-
-    useEffect(() => {
-        extractSchemaRef.current = extractSchema;
-    }, [extractSchema]);
 
     useEffect(() => {
         const handleScript = () => {
@@ -267,18 +218,6 @@ const SchemaDesignerStateProvider: React.FC<SchemaDesignerProviderProps> = ({ ch
             extractSchema,
         });
     }, [isInitialized, extensionRpc, extractSchema]);
-
-    useEffect(() => {
-        registerSchemaDesignerDabToolHandlers({
-            extensionRpc,
-            isInitializedRef,
-            getCurrentDabConfig: () => dabConfigRef.current,
-            getCurrentSchemaTables: () => extractSchemaRef.current().tables,
-            commitDabConfig: (config) => {
-                setDabConfig(config);
-            },
-        });
-    }, [extensionRpc]);
 
     const initializeSchemaDesigner = async () => {
         try {
@@ -708,277 +647,8 @@ const SchemaDesignerStateProvider: React.FC<SchemaDesignerProviderProps> = ({ ch
         }, 10);
     }
 
-    // DAB functions
-    const initializeDabConfig = useCallback(() => {
-        const schema = extractSchema();
-        const config = Dab.createDefaultConfig(schema.tables);
-        setDabConfig(config);
-    }, [reactFlow]);
-
-    const syncDabConfigWithSchema = useCallback(() => {
-        if (!dabConfig) {
-            return;
-        }
-
-        const schema = extractSchema();
-        const currentTableIds = new Set(schema.tables.map((t) => t.id));
-        const existingEntityIds = new Set(dabConfig.entities.map((e) => e.id));
-
-        // Find new tables that need to be added
-        const newTables = schema.tables.filter((t) => !existingEntityIds.has(t.id));
-
-        // Filter out entities for tables that no longer exist
-        const updatedEntities = dabConfig.entities.filter((e) => currentTableIds.has(e.id));
-
-        // Add new tables with default config
-        const newEntities = newTables.map((t) => Dab.createDefaultEntityConfig(t));
-
-        // Only update if there are changes
-        if (newEntities.length > 0 || updatedEntities.length !== dabConfig.entities.length) {
-            setDabConfig({
-                ...dabConfig,
-                entities: [...updatedEntities, ...newEntities],
-            });
-        }
-    }, [dabConfig, reactFlow]);
-
-    const updateDabApiTypes = useCallback((apiTypes: Dab.ApiType[]) => {
-        setDabConfig((prev) => {
-            if (!prev) {
-                return prev;
-            }
-            return {
-                ...prev,
-                apiTypes,
-            };
-        });
-    }, []);
-
-    const toggleDabEntity = useCallback((entityId: string, isEnabled: boolean) => {
-        setDabConfig((prev) => {
-            if (!prev) {
-                return prev;
-            }
-            return {
-                ...prev,
-                entities: prev.entities.map((e) => (e.id === entityId ? { ...e, isEnabled } : e)),
-            };
-        });
-    }, []);
-
-    const toggleDabEntityAction = useCallback(
-        (entityId: string, action: Dab.EntityAction, isEnabled: boolean) => {
-            setDabConfig((prev) => {
-                if (!prev) {
-                    return prev;
-                }
-                return {
-                    ...prev,
-                    entities: prev.entities.map((e) => {
-                        if (e.id !== entityId) {
-                            return e;
-                        }
-                        const enabledActions = isEnabled
-                            ? [...e.enabledActions, action]
-                            : e.enabledActions.filter((a) => a !== action);
-                        return { ...e, enabledActions };
-                    }),
-                };
-            });
-        },
-        [],
-    );
-
-    const updateDabEntitySettings = useCallback(
-        (entityId: string, settings: Dab.EntityAdvancedSettings) => {
-            setDabConfig((prev) => {
-                if (!prev) {
-                    return prev;
-                }
-                return {
-                    ...prev,
-                    entities: prev.entities.map((e) =>
-                        e.id === entityId ? { ...e, advancedSettings: settings } : e,
-                    ),
-                };
-            });
-        },
-        [],
-    );
-
-    const generateDabConfig = useCallback(async () => {
-        if (!dabConfig) {
-            return;
-        }
-        const response = await extensionRpc.sendRequest(Dab.GenerateConfigRequest.type, {
-            config: dabConfig,
-        });
-        if (response.success) {
-            setDabConfigContent(response.configContent);
-            setDabConfigRequestId((id) => id + 1);
-        }
-    }, [dabConfig, extensionRpc]);
-
-    const openDabConfigInEditor = useCallback(
-        (configContent: string) => {
-            void extensionRpc.sendNotification(Dab.OpenConfigInEditorNotification.type, {
-                configContent,
-            });
-        },
-        [extensionRpc],
-    );
-
     const dabEnabled = useSchemaDesignerSelector((s) => s?.enableDAB);
     const isDabEnabled = () => dabEnabled ?? false;
-
-    // DAB Deployment functions
-    const openDabDeploymentDialog = useCallback(() => {
-        setDabDeploymentState((prev) => ({
-            ...prev,
-            isDialogOpen: true,
-            dialogStep: Dab.DabDeploymentDialogStep.Confirmation,
-        }));
-    }, []);
-
-    const closeDabDeploymentDialog = useCallback(() => {
-        setDabDeploymentState((prev) => ({
-            ...prev,
-            isDialogOpen: false,
-        }));
-    }, []);
-
-    const setDabDeploymentDialogStep = useCallback((step: Dab.DabDeploymentDialogStep) => {
-        setDabDeploymentState((prev) => ({
-            ...prev,
-            dialogStep: step,
-        }));
-    }, []);
-
-    const updateDabDeploymentParams = useCallback((params: Partial<Dab.DabDeploymentParams>) => {
-        setDabDeploymentState((prev) => ({
-            ...prev,
-            params: {
-                ...prev.params,
-                ...params,
-            },
-        }));
-    }, []);
-
-    const validateDabDeploymentParams = useCallback(
-        async (
-            containerName: string,
-            port: number,
-        ): Promise<Dab.ValidateDeploymentParamsResponse> => {
-            return extensionRpc.sendRequest(Dab.ValidateDeploymentParamsRequest.type, {
-                containerName,
-                port,
-            });
-        },
-        [extensionRpc],
-    );
-
-    const updateDeploymentStepStatus = useCallback(
-        (
-            step: Dab.DabDeploymentStepOrder,
-            status: ApiStatus,
-            message?: string,
-            fullErrorText?: string,
-            errorLink?: string,
-            errorLinkText?: string,
-        ) => {
-            setDabDeploymentState((prev) => ({
-                ...prev,
-                stepStatuses: prev.stepStatuses.map((s) =>
-                    s.step === step
-                        ? { ...s, status, message, fullErrorText, errorLink, errorLinkText }
-                        : s,
-                ),
-            }));
-        },
-        [],
-    );
-
-    const runDabDeploymentStep = useCallback(
-        async (step: Dab.DabDeploymentStepOrder) => {
-            // Mark step as running
-            updateDeploymentStepStatus(step, ApiStatus.Loading);
-
-            // For container start step, verify DAB config is available
-            if (step === Dab.DabDeploymentStepOrder.startContainer && !dabConfig) {
-                updateDeploymentStepStatus(
-                    step,
-                    ApiStatus.Error,
-                    "DAB configuration is not available.",
-                );
-                return;
-            }
-
-            // Run the step via extension (extension will generate config content as needed)
-            const response = await extensionRpc.sendRequest(Dab.RunDeploymentStepRequest.type, {
-                step,
-                params: dabDeploymentState.params,
-                config: dabConfig ?? undefined,
-            });
-
-            if (response.success) {
-                // Update step status to completed and advance to next step
-                setDabDeploymentState((prev) => {
-                    const updatedStatuses = prev.stepStatuses.map((s) =>
-                        s.step === step ? { ...s, status: ApiStatus.Loaded } : s,
-                    );
-
-                    // If this was the last step, set completion state
-                    if (step === Dab.DabDeploymentStepOrder.checkContainer) {
-                        return {
-                            ...prev,
-                            stepStatuses: updatedStatuses,
-                            currentDeploymentStep: step + 1,
-                            isDeploying: false,
-                            apiUrl: response.apiUrl,
-                            dialogStep: Dab.DabDeploymentDialogStep.Complete,
-                        };
-                    }
-
-                    return {
-                        ...prev,
-                        stepStatuses: updatedStatuses,
-                        currentDeploymentStep: step + 1,
-                    };
-                });
-            } else {
-                updateDeploymentStepStatus(
-                    step,
-                    ApiStatus.Error,
-                    response.error,
-                    response.fullErrorText,
-                    response.errorLink,
-                    response.errorLinkText,
-                );
-            }
-        },
-        [dabConfig, dabDeploymentState.params, extensionRpc, updateDeploymentStepStatus],
-    );
-
-    const resetDabDeploymentState = useCallback(() => {
-        setDabDeploymentState(Dab.createDefaultDeploymentState());
-    }, []);
-
-    const retryDabDeploymentSteps = useCallback(() => {
-        // Reset only deployment steps (pullImage, startContainer, checkContainer)
-        // while keeping prerequisite steps as completed
-        setDabDeploymentState((prev) => ({
-            ...prev,
-            currentDeploymentStep: Dab.DabDeploymentStepOrder.pullImage,
-            stepStatuses: prev.stepStatuses.map((s) => {
-                if (s.step >= Dab.DabDeploymentStepOrder.pullImage) {
-                    return { ...s, status: ApiStatus.NotStarted, message: undefined };
-                }
-                return s;
-            }),
-            error: undefined,
-            apiUrl: undefined,
-        }));
-    }, []);
 
     return (
         <SchemaDesignerContext.Provider
@@ -1021,31 +691,7 @@ const SchemaDesignerStateProvider: React.FC<SchemaDesignerProviderProps> = ({ ch
                 baselineRevision,
                 schemaRevision,
                 notifySchemaChanged,
-                // DAB state
-                dabConfig,
                 isDabEnabled,
-                initializeDabConfig,
-                syncDabConfigWithSchema,
-                updateDabApiTypes,
-                toggleDabEntity,
-                toggleDabEntityAction,
-                updateDabEntitySettings,
-                dabSchemaFilter,
-                setDabSchemaFilter,
-                dabConfigContent,
-                dabConfigRequestId,
-                generateDabConfig,
-                openDabConfigInEditor,
-                // DAB Deployment state
-                dabDeploymentState,
-                openDabDeploymentDialog,
-                closeDabDeploymentDialog,
-                setDabDeploymentDialogStep,
-                updateDabDeploymentParams,
-                validateDabDeploymentParams,
-                runDabDeploymentStep,
-                resetDabDeploymentState,
-                retryDabDeploymentSteps,
             }}>
             {children}
         </SchemaDesignerContext.Provider>
